Cleanup of USB Driver
Due to the fix with compiler issue #2 and the update of the memory pointer libraries to support PICs that do not have the POSTINC function, all changes done in the USB driver in 2018 could be reverted. So the driver is more or less restored to its original state but now it supports both PICs with and without the POSTINC function.  In this update the TABS used in the JAL files where replaced by spaces.